are coming closer to
link>
http://www.fsf.org/blogs/licensing/the-sharks-move-in-lobbyists-pushing-forward-on-tpp-agreements
</link>
<description>
The latest leaked draft of the TPP reveals that the countries involved in the negotiations are coming closer to acceptance of a whole host of problematic agreements.
</description>
<content:encoded xmlns:content="http://purl.org/rss/1.0/modules/content/">
<![CDATA[
<p>On October 16th, <a href="https://www.wikileaks.org/tpp-ip2/">WikiLeaks</a> released an updated draft of the Trans-Pacific Partnership (TPP) Strategic Partnership Agreement chapter on copyright, patent and other proprietary interests. A previous draft had been released last year. If you aren't fa
]]>
<![CDATA[
miliar with TPP, it is a multinational trade-agreement that is being developed through a series of secret negotiations that when enacted will have a vast effect on civil liberties, including the ability of users all around the world to enjoy software freedom. </p> <p>We have been <a href="https://www.fsf.org/blogs/community/2012-melbourne-tpp-report">following</a> and <a href="https://www.fsf.org/blogs/community/fsf-speaks-against-tpp">opposing</a> these negotiations both in person and online for many years now. We <a href="https://www.fsf.org/blogs/community/as-free-software-users-we-need-to-speak-out-against-the-tpp">wrote earlier this year</a> about the dangers posed by these secret negotiations. This latest leak reveals that the countries involved in the TPP negotiations are coming closer to acceptance of a whole host of problematic agreements:</p> <ul> <li><strong>Penalizing the circumvention of digital restriction management, even for non-infringing uses</strong>. Previous leaked versions of the negotiations were poised to repeat the <a href="http://www.defectivebydesign.org/2014-cellphone-unlocking-bill">miserable failure of the United State's DMCA anti-circumvention exemption regime</a>. While the current draft has moved away from implementing this broken system elsewhere, it still leaves in place a system where users can face penalties for circumventing digital restrictions management. Worse still, these penalties can apply even when the circumvention is done for non-infringing uses.</li> <li><strong>Perpetuating indefinite terms of copyright</strong>. The current draft shows the parties solidifying agreement around extending the term of copyright restriction, potentially up to life of the author plus 100 years. The U.S. has extended the term of copyright several times over the past decades, and accepting the maximum proposed change would once again <a href="https://www.gnu.org/philosophy/misinterpreting-copyright.html">break the basic bargain of the copyright system</a>. This term would push perpetual copyright into other countries as well.</li> <li><strong>Failing to block the patenting of software</strong>. The current draft once again leaves open the possibility of software coming under the heading of patentable subject matter. The disaster of software patents in countries where they already exist is one of the <a href="http://endsoftpatents.org/">greatest threats to user freedom</a>, and a regime that does not block the patenting of software will only expand this problem elsewhere. Most troubling is the proposal by Mexico specifically stating that signing parties should be able to exclude software from patentable subject matter. This would be great news, if not for the fact that they are the only country signed onto the proposal.</li> </ul> <p>These are just a few of the problems with the subject matter of the negotiations, not to mention the over-arching problem of the process being hidden from the public. We have asked for your help in the past in opposing TPP, but the fight is still not over. Here's what you can do to help:</p> <ul> <li>Write to your representatives telling them you oppose TPP.</li> <li>Support our work representing you in this area by <a href="https://crm.fsf.org/join">joining as a member</a> or <a href="https://donate.fsf.org">making a donation</a>.</li> </ul>
]]>
</content:encoded>
<dc:publisher>No publisher</dc:publisher>
<dc:creator/>
<dc:rights/>
<dc:date>2014-11-24T22:30:00Z</dc:date>
<dc:type>Blog Entry</dc:type>
</item>
</rdf:RDF>